+++ src/main/java/br/com/gustavo/springsecurity/controller/dto/FeedDTO.java
@@ -2,5 +2,5 @@ package br.com.gustavo.springsecurity.controller.dto;
 
 import java.util.List;
 
-public record FeedDTO(List<FeedItemDTO> feedItems, int page, int pageSize, int totalPages, int totalElements) {
+public record FeedDTO(List<FeedItemDTO> feedItems, int page, int pageSize, int totalPages, Long totalElements) {
 }
